requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14,<3.16
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9.9,<3.0

# Environment & Config
python-dotenv>=1.0.0,<2.0

# Static Files (Production)
whitenoise>=6.6.0,<7.0

# Production Server
gunicorn>=21.0.0,<22.0
